pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-ult-assembly"
version = "0.1.0"
description = "Scaffold workspace for assembly."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.6",
    "pyyaml>=6.0",
]

[project.scripts]
assembly = "assembly.cli.main:entrypoint"

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
